Split callback_fn into submodules
Simplifies managing visibility and conditional compilation based on
feature flags.
